git-svn-id: http://svn.boost.org/svn/boost/branches/release@64101 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/archive/basic_archive.hpp
@@ -74,7 +74,7 @@ BOOST_ARCHIVE_VERSION();
 
 class version_type {
 private:
-    typedef uint_least8_t base_type;
+    typedef unsigned int base_type;
     base_type t;
     version_type(): t(0) {};
 public:
